pyproject.toml
[project]
name = "task-manager-api"
version = "0.1.0"
description = ""
authors = [
    {name = "KamranYsupov",email = "kamrankek48@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "sqlalchemy[asyncio] (>=2.0.40,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "dependency-injector (>=4.46.0,<5.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "orjson (>=3.10.18,<4.0.0)",
    "loguru (==0.6.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest[asyncio] (>=8.4.1,<9.0.0)",
    "pytest-asyncio (>=1.1.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "q", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
